requirements.txt
tqdm==4.41.1
matplotlib==3.1.2
pandas==0.25.3
torch==1.3.1
torchvision==0.4.2
requests==2.18.4
Pillow==7.0.0
